Revert "counts"
This reverts commit 0a4f3851a635eb84c632119a2e5ea67b050ce7ed.

+++ include/seqan/kmer/kmer_ibf.h
@@ -303,7 +303,7 @@ public:
     template<typename TInt>
     inline void select(std::vector<bool> & selected, TString const & text, TInt && threshold)
     {
-        std::vector<uint32_t> counts(noOfBins, 0);
+        std::vector<uint16_t> counts(noOfBins, 0);
         select(counts, text);
         for(uint16_t binNo=0; binNo < noOfBins; ++binNo)
         {
